requirements.txt
python-dotenv
flask
slackeventsapi
requests
slack
slackclient
slack_sdk